refactor: move currency dialog into UI
