+++ src/currencymanager.h
@@ -3,21 +3,19 @@
 
 #pragma once
 
-#include <QCheckBox>
-#include <QDialog>
-#include <QLabel>
 #include <QObject>
 #include <QString>
-#include <QWidget>
 
-#include "buyoutmanager.h"
+#include <array>
+#include <memory>
+#include <vector>
 
-class QDoubleSpinBox;
+#include "currency.h"
+
 class QSettings;
-class QVBoxLayout;
 
-class CurrencyManager;
 class DataStore;
+class Item;
 class ItemsManager;
 
 struct CurrencyRatio
@@ -54,54 +52,6 @@ struct CurrencyItem
         , exalt(curr, Currency::CURRENCY_EXALTED_ORB, exalt_ratio, 1)
         , chaos(curr, Currency::CURRENCY_CHAOS_ORB, chaos_ratio, 1) {};
 };
-struct CurrencyLabels
-{
-    QLabel *name;
-    QLabel *count;
-    QLabel *chaos_ratio;
-    QLabel *chaos_value;
-    QLabel *exalt_ratio;
-    QLabel *exalt_value;
-    QLabel *exalt_total;
-    QLabel *chaos_total;
-    QLabel *wisdom_total;
-    CurrencyLabels()
-    {
-        name = new QLabel("Name");
-        count = new QLabel("Count");
-        chaos_ratio = new QLabel("Amount a chaos Orb can buy");
-        chaos_value = new QLabel("Value in Chaos Orb");
-        exalt_ratio = new QLabel("Amount an Exalted Orb can buy");
-        exalt_value = new QLabel("Value in Exalted Orb");
-        exalt_total = new QLabel("Total Exalted Orbs");
-        chaos_total = new QLabel("Total Chaos Orbs");
-        wisdom_total = new QLabel("Total Scrolls of Wisdom");
-    }
-};
-class CurrencyDialog;
-class CurrencyWidget : public QWidget
-{
-    Q_OBJECT
-public slots:
-    void Update();
-    void UpdateVisual(bool show_chaos, bool show_exalt);
-
-public:
-    CurrencyWidget(std::shared_ptr<CurrencyItem> currency);
-    bool IsNone() const { return m_currency->currency.type == Currency::CURRENCY_NONE; }
-    //Visual stuff
-    QLabel *name;
-    QLabel *count;
-    QDoubleSpinBox *chaos_ratio;
-    QDoubleSpinBox *chaos_value;
-    QDoubleSpinBox *exalt_ratio;
-    QDoubleSpinBox *exalt_value;
-
-private:
-    //Data
-    std::shared_ptr<CurrencyItem> m_currency;
-};
-
 constexpr std::array<const char *, 5> CurrencyForWisdom({{"Scroll of Wisdom",
                                                           "Portal Scroll",
                                                           "Armourer's Scrap",
@@ -110,35 +60,6 @@ constexpr std::array<const char *, 5> CurrencyForWisdom({{"Scroll of Wisdom",
 
 constexpr std::array<int, 5> CurrencyWisdomValue({{1, 1, 2, 4, 4}});
 
-class CurrencyDialog : public QDialog
-{
-    Q_OBJECT
-public:
-    CurrencyDialog(CurrencyManager &manager, bool show_chaos, bool show_exalt);
-    bool ShowChaos() const { return m_show_chaos->isChecked(); }
-    bool ShowExalt() const { return m_show_exalt->isChecked(); }
-
-public slots:
-    void Update();
-    void UpdateVisual();
-    void UpdateVisibility(bool show_chaos, bool show_exalt);
-    void UpdateTotalValue();
-
-private:
-    CurrencyManager &m_currency_manager;
-    std::vector<CurrencyWidget *> m_currencies_widgets;
-    CurrencyLabels *m_headers;
-    QVBoxLayout *m_layout;
-    QLabel *m_total_exalt_value;
-    QLabel *m_total_chaos_value;
-    QLabel *m_total_wisdom_value;
-    QCheckBox *m_show_chaos;
-    QCheckBox *m_show_exalt;
-    QFrame *m_separator;
-    QVBoxLayout *GenerateLayout(bool show_chaos, bool show_exalt);
-    void UpdateTotalWisdomValue();
-};
-
 class CurrencyManager : public QObject
 {
     Q_OBJECT
@@ -146,7 +67,7 @@ public:
     explicit CurrencyManager(QSettings &settings, DataStore &datastore, ItemsManager &items_manager);
     ~CurrencyManager();
     void ClearCurrency();
-    // Called in itemmanagerworker::ParseItem
+    // Called from CurrencyManager::Update.
     void ParseSingleItem(const Item &item);
     //void UpdateBaseValue(int ind, double value);
     const std::vector<std::shared_ptr<CurrencyItem>> &currencies() const { return m_currencies; }
@@ -154,10 +75,12 @@ public:
     double TotalChaosValue();
     int TotalWisdomValue();
     void Update();
+    void ExportCurrency(const QString &file_name);
+
+signals:
+    void Updated();
 
 public slots:
-    void DisplayCurrency();
-    void ExportCurrency();
     void SaveCurrencyValue();
 
 private:
@@ -168,7 +91,6 @@ private:
     std::vector<std::shared_ptr<CurrencyItem>> m_currencies;
     // We only need the "count" of a CurrencyItem so int will be enough
     std::vector<int> m_wisdoms;
-    std::shared_ptr<CurrencyDialog> m_dialog;
     // Used only the first time we launch the app
     void FirstInitCurrency();
     //Migrate from old storage (csv-like serializing) to new one (using json)
